refactor: 실행 결과 도메인 구현
